media: Add InputHandler
InputHandler controls decoder and streambuffer which were handled by
InputDataSource.
Related method and variables are moved to InputHandler.
