+++ framework/include/media/InputDataSource.h
@@ -29,27 +29,11 @@
 #ifndef __MEDIA_INPUTDATASOURCE_H
 #define __MEDIA_INPUTDATASOURCE_H
 
-#include <fstream>
 #include <memory>
-#include <pthread.h>
 #include <media/DataSource.h>
-#include <media/BufferObserverInterface.h>
 
 namespace media {
-class Decoder;
-class MediaPlayerImpl;
-
-typedef enum buffer_state_e {
-	BUFFER_STATE_EMPTY,
-	BUFFER_STATE_BUFFERING,
-	BUFFER_STATE_BUFFERED,
-	BUFFER_STATE_FULL,
-} buffer_state_t;
-
 namespace stream {
-class StreamBuffer;
-class StreamBufferReader;
-class StreamBufferWriter;
 
 /**
  * @class
@@ -59,7 +43,6 @@ class StreamBufferWriter;
  */
 class InputDataSource
 	: public DataSource
-	, public BufferObserverInterface
 {
 public:
 	/**
@@ -92,7 +75,7 @@ public:
 	 * @details @b #include <media/InputDataSource.h>
 	 * @since TizenRT v2.0 PRE
 	 */
-	virtual ssize_t read(unsigned char *buf, size_t size);
+	virtual ssize_t read(unsigned char *buf, size_t size) = 0;
 
 	/**
 	 * @brief Set the offset
@@ -113,59 +96,11 @@ public:
 	 */
 	virtual int readAt(long offset, int origin, unsigned char *buf, size_t size);
 
-	/**
-	 * @brief Register current player to get data souce state and other informations.
-	 * @details @b #include <media/InputDataSource.h>
-	 * @since TizenRT v2.0 PRE
-	 */
-	void setPlayer(std::shared_ptr<MediaPlayerImpl> mp) { mPlayer = mp; }
-
-	/* BufferObserverInterface
-	 * Override these methods in derived class if needed.
-	 */
-	virtual void onBufferOverrun() override;
-	virtual void onBufferUnderrun() override;
-	virtual void onBufferUpdated(ssize_t change, size_t current) override;
-
-protected:
-	bool start();
-	bool stop();
-	virtual ssize_t onStreamBufferWritable() = 0;
-
-protected:
 	void setAudioType(audio_type_t audioType);
 	audio_type_t getAudioType();
-	bool registerDecoder(audio_type_t audioType, unsigned int channels, unsigned int sampleRate);
-	void unregisterDecoder();
-	size_t getDecodeFrames(unsigned char *buf, size_t *size);
-
-	void setStreamBuffer(std::shared_ptr<StreamBuffer>);
-	std::shared_ptr<StreamBuffer> getStreamBuffer() { return mStreamBuffer; }
-	std::shared_ptr<StreamBufferWriter> getBufferWriter() { return mBufferWriter; }
-	void setBufferState(buffer_state_t state);
-	std::shared_ptr<MediaPlayerImpl> getPlayer() { return mPlayer.lock(); }
-	void createWorker();
-	void destoryWorker();
-	ssize_t writeToStreamBuffer(unsigned char *buf, size_t size);
 
 private:
 	audio_type_t mAudioType;
-	std::shared_ptr<Decoder> mDecoder;
-	std::shared_ptr<StreamBuffer> mStreamBuffer;
-	std::shared_ptr<StreamBufferReader> mBufferReader;
-	std::shared_ptr<StreamBufferWriter> mBufferWriter;
-	std::weak_ptr<MediaPlayerImpl> mPlayer;
-
-	buffer_state_t mState;
-	size_t mTotalBytes;
-	bool mIsWorkerAlive;
-	pthread_t mWorker;
-	std::mutex mMutex;
-	std::condition_variable mCondv;
-
-	void sleepWorker();
-	void wakenWorker();
-	static void *workerMain(void *arg);
 };
 
 } // namespace stream
