+++ framework/src/media/Make.defs
@@ -44,7 +44,9 @@ VPATH += :src/media/codecs
 endif
 
 ifeq ($(CONFIG_MEDIA_PLAYER), y)
-CXXSRCS += MediaPlayer.cpp InputDataSource.cpp FileInputDataSource.cpp PlayerWorker.cpp MediaPlayerImpl.cpp BufferInputDataSource.cpp PlayerObserverWorker.cpp
+CXXSRCS += MediaPlayer.cpp PlayerWorker.cpp MediaPlayerImpl.cpp PlayerObserverWorker.cpp
+CXXSRCS += InputHandler.cpp
+CXXSRCS += InputDataSource.cpp FileInputDataSource.cpp
 CXXSRCS += Decoder.cpp audio_decoder.cpp
 endif
